ah, D'oh!
We fork when we get a tcp request; then when we update the database it
might get corrupted and we crash and can't start because the database is
foobared.

as long as we use geod anyway it's pretty cheap to lookup again, so let's
use Tie::Cache::LRU for now instead of DB_File.